+++ src/output/p_ac3.cpp
@@ -172,8 +172,8 @@ ac3_packetizer_c::process(packet_cptr packet) {
 
 void
 ac3_packetizer_c::dump_debug_info() {
-  mxdebug("ac3_packetizer_c: queue: %d; buffer size: %d\n",
-          packet_queue.size(), byte_buffer.get_size());
+  mxdebug("ac3_packetizer_c: queue: %u; buffer size: %d\n",
+          (unsigned int)packet_queue.size(), byte_buffer.get_size());
 }
 
 connection_result_e
