+++ src/output/p_dts.cpp
@@ -202,7 +202,7 @@ dts_packetizer_c::process(packet_cptr packet) {
 
 void
 dts_packetizer_c::dump_debug_info() {
-  mxdebug("dts_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("dts_packetizer_c: queue: %u\n", (unsigned int)packet_queue.size());
 }
 
 connection_result_e
