+++ src/output/p_flac.cpp
@@ -109,7 +109,7 @@ flac_packetizer_c::process(packet_cptr packet) {
 
 void
 flac_packetizer_c::dump_debug_info() {
-  mxdebug("flac_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("flac_packetizer_c: queue: %u\n", (unsigned int)packet_queue.size());
 }
 
 connection_result_e
