+++ src/output/p_textsubs.cpp
@@ -144,7 +144,8 @@ textsubs_packetizer_c::process(packet_cptr packet) {
 
 void
 textsubs_packetizer_c::dump_debug_info() {
-  mxdebug("textsubs_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("textsubs_packetizer_c: queue: %u\n",
+          (unsigned int)packet_queue.size());
 }
 
 connection_result_e
