+++ src/output/p_tta.cpp
@@ -76,7 +76,7 @@ tta_packetizer_c::process(packet_cptr packet) {
 
 void
 tta_packetizer_c::dump_debug_info() {
-  mxdebug("tta_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("tta_packetizer_c: queue: %u\n", (unsigned int)packet_queue.size());
 }
 
 connection_result_e
