+++ src/output/p_vobsub.cpp
@@ -70,7 +70,8 @@ vobsub_packetizer_c::process(packet_cptr packet) {
 
 void
 vobsub_packetizer_c::dump_debug_info() {
-  mxdebug("vobsub_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("vobsub_packetizer_c: queue: %u\n",
+          (unsigned int)packet_queue.size());
 }
 
 connection_result_e
