+++ src/output/p_vorbis.cpp
@@ -226,7 +226,8 @@ vorbis_packetizer_c::process(packet_cptr packet) {
 
 void
 vorbis_packetizer_c::dump_debug_info() {
-  mxdebug("vorbis_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("vorbis_packetizer_c: queue: %u\n",
+          (unsigned int)packet_queue.size());
 }
 
 connection_result_e
